refactored AutoGluon Adapter
