+++ adapters/AutoGluon/AutoGluonServer.py
@@ -1,23 +1,48 @@
+import grpc
 import os
-
 import logging
 import json
 import time
-from concurrent import futures
+import shutil
 
-import grpc
+from autogluon.tabular import TabularPredictor
 
 import Adapter_pb2
 import Adapter_pb2_grpc
 
+from concurrent import futures
 from JsonUtil import get_config_property
 
 from AdapterUtils import *
 
-
-
-
-
+def GetMetaInformations(config_json):
+    working_dir = os.path.join(get_config_property("output-path"), "working_dir")
+    shutil.unpack_archive(os.path.join(get_config_property("output-path"),
+        str(config_json["session_id"]),
+        get_config_property("export-zip-file-name") + ".zip"),
+        working_dir,
+        "zip")
+    # extract additional information from automl
+    automl = TabularPredictor.load(os.path.join(os.path.join(get_config_property("output-path"), "working_dir"), 'model_gluon.gluon'))
+    automl_info = automl._learner.get_info(include_model_info=True)
+    librarylist = set()
+    model = automl_info['best_model']
+    for model_info in automl_info['model_info']:
+        if model_info == model:
+            pass
+        elif model_info in ('LightGBM', 'LightGBMXT'):
+            librarylist.add("lightgbm")
+        elif model_info == 'XGBoost':
+            librarylist.add("xgboost")
+        elif model_info == 'CatBoost':
+            librarylist.add("catboost")
+        elif model_info == 'NeuralNetFastAI':
+            librarylist.add("pytorch")
+        else:
+            librarylist.add("sklearn")
+    library = " + ".join(librarylist)
+    shutil.rmtree(working_dir)
+    return library, model
 
 class AdapterServiceServicer(Adapter_pb2_grpc.AdapterServiceServicer):
     """
@@ -42,16 +67,18 @@ class AdapterServiceServicer(Adapter_pb2_grpc.AdapterServiceServicer):
             yield from capture_process_output(process, start_time)
             generate_script(config_json)
             output_json = zip_script(config_json["session_id"])
-            library = "neural network"
-            model = "keras"
-            test_score, prediction_time = evaluate(config_json, job_file_location)
+
+
+            library, model = GetMetaInformations(config_json)
+            test_score, prediction_time= evaluate(config_json, job_file_location)
             response = yield from get_response(output_json, start_time, test_score, prediction_time, library, model)
+
             print(f'{get_config_property("adapter-name")} job finished')
             return response
 
         except Exception as e:
             return get_except_response(context, e)
-
+    
     def TestAdapter(self, request, context):
         try:
             # saving AutoML configuration JSON
@@ -69,7 +96,6 @@ class AdapterServiceServicer(Adapter_pb2_grpc.AdapterServiceServicer):
 
         except Exception as e:
             return get_except_response(context, e)
-
 
 def serve():
     """
