+++ adapters/AutoGluon/requirements.txt
@@ -1,23 +1,3 @@
-autogluon==0.3.1
-numpy>=1.16.0
-scipy>=1.3.3
-cython
-tornado>=5.0.1
-requests
-matplotlib
-tqdm>=4.38.0
-paramiko>=2.5.0
-dask==2.6.0
-cryptography>=2.8
-distributed==2.6.0
-gluoncv>=0.5.0
-gluonnlp==0.8.1
-graphviz
-scikit-optimize
-catboost
-boto3
-psutil>=5.0.0
-scikit-learn>=0.20.0
-mxnet
-mxnet_cu101
-grpcio==1.34.1
+autogluon
+grpcio
+dill
