+++ adapters/Utils/AutoMLs/AbstractAdapter.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+import dill
 
 
 class AbstractAdapter(ABC):
@@ -26,3 +27,13 @@ class AbstractAdapter(ABC):
     def start(self):
         """Execute the ML task"""
         pass
+
+    def __export_model(self, model, file_name):
+        """
+        Export the generated ML model to disk
+        ---
+        Parameter:
+        1. generate ML model
+        """
+        with open(os.path.join(get_config_property('output-path'), 'tmp', file_name), 'wb+') as file:
+            dill.dump(model, file)
